Bullet emitter rotates to mouse
I gave it some thought. Player should probably be able to control the exact direction of the gun. If it's locked to the axis in a game where AI can approach from any angle, it would get annoying to use the last resort weapon.

The bullet emitter now rotates to face the cursor.
Placeholder art easier to interpret.